Created app for sulten and moved views, models, admin and urls
