+++ backend/samfundet/admin.py
@@ -22,26 +22,19 @@ from .models.recruitment import (
 from .models.general import (
     Tag,
     User,
-    Menu,
     Gang,
     Venue,
-    Table,
     Image,
     Campus,
     Profile,
-    Booking,
     Infobox,
-    MenuItem,
     GangType,
     TextItem,
     KeyValue,
     BlogPost,
     Organization,
-    Reservation,
     ClosedPeriod,
     Saksdokument,
-    FoodCategory,
-    FoodPreference,
     UserPreference,
     InformationPage,
 )
@@ -341,88 +334,6 @@ class BlogPostAdmin(CustomBaseAdmin):
     list_select_related = True
 
 
-@admin.register(Table)
-class TableAdmin(CustomBaseAdmin):
-    # ordering = []
-    sortable_by = ['id', 'name_nb', 'name_en', 'seating', 'created_at', 'updated_at']
-    # list_filter = []
-    list_display = ['id', '__str__', 'name_nb', 'name_en', 'seating', 'created_at', 'updated_at']
-    search_fields = ['id', 'name_nb', 'name_en']
-    # filter_horizontal = []
-    list_display_links = ['id', '__str__']
-    # autocomplete_fields = []
-    list_select_related = True
-
-
-@admin.register(Reservation)
-class ReservationAdmin(CustomBaseAdmin):
-    # ordering = []
-    sortable_by = ['id', 'name', 'email', 'phonenumber']
-    # list_filter = []
-    list_display = ['id', '__str__', 'name', 'email', 'phonenumber']
-    search_fields = ['id', 'name', 'email', 'phonenumber']
-    # filter_horizontal = []
-    list_display_links = ['id', '__str__']
-    # autocomplete_fields = []
-    list_select_related = True
-
-
-@admin.register(Menu)
-class MenuAdmin(CustomBaseAdmin):
-    # ordering = []
-    sortable_by = ['id', 'name_nb', 'name_en', 'created_at', 'updated_at']
-    # list_filter = []
-    list_display = ['id', '__str__', 'name_nb', 'name_en', 'menu_item_count', 'created_at', 'updated_at']
-    search_fields = ['id', 'name_nb', 'name_en']
-    filter_horizontal = ['menu_items']
-    list_display_links = ['id', '__str__']
-    # autocomplete_fields = []
-    list_select_related = True
-
-    def menu_item_count(self, obj: Menu) -> int:
-        n: int = obj.menu_items.all().count()
-        return n
-
-
-@admin.register(MenuItem)
-class MenuItemAdmin(CustomBaseAdmin):
-    # ordering = []
-    sortable_by = ['id', 'name_nb', 'name_en', 'price', 'price_member', 'order', 'created_at', 'updated_at']
-    # list_filter = []
-    list_display = ['id', '__str__', 'name_nb', 'name_en', 'price', 'price_member', 'order', 'created_at', 'updated_at']
-    search_fields = ['id', 'name_nb', 'name_en']
-    filter_horizontal = ['food_preferences']
-    list_display_links = ['id', '__str__']
-    # autocomplete_fields = []
-    list_select_related = True
-
-
-@admin.register(FoodCategory)
-class FoodCategoryAdmin(CustomBaseAdmin):
-    # ordering = []
-    sortable_by = ['id', 'name_nb', 'name_en', 'order', 'created_at', 'updated_at']
-    # list_filter = []
-    list_display = ['id', '__str__', 'name_nb', 'name_en', 'order', 'created_at', 'updated_at']
-    search_fields = ['id', 'name_nb', 'name_en']
-    # filter_horizontal = []
-    list_display_links = ['id', '__str__']
-    # autocomplete_fields = []
-    list_select_related = True
-
-
-@admin.register(FoodPreference)
-class FoodPreferenceAdmin(CustomBaseAdmin):
-    # ordering = []
-    sortable_by = ['id', 'name_nb', 'name_en', 'created_at', 'updated_at']
-    # list_filter = []
-    list_display = ['id', '__str__', 'name_nb', 'name_en', 'created_at', 'updated_at']
-    search_fields = ['id', 'name_nb', 'name_en']
-    # filter_horizontal = []
-    list_display_links = ['id', '__str__']
-    # autocomplete_fields = []
-    list_select_related = True
-
-
 @admin.register(Saksdokument)
 class SaksdokumentAdmin(CustomBaseAdmin):
     # ordering = []
@@ -433,19 +344,6 @@ class SaksdokumentAdmin(CustomBaseAdmin):
     # filter_horizontal = []
     list_display_links = ['id', '__str__']
     # autocomplete_fields = []
-
-
-@admin.register(Booking)
-class BookingAdmin(CustomBaseAdmin):
-    # ordering = []
-    # list_filter = []
-    list_display = ['id', '__str__', 'name', 'get_duration', 'table_count', 'created_at', 'updated_at']
-    _user_search_fields = UserAdmin.custom_search_fields(prefix='user')
-    search_fields = ['id', 'name', *_user_search_fields]
-    filter_horizontal = ['tables']
-    list_display_links = ['id', '__str__']
-    autocomplete_fields = ['user']
-    list_select_related = True
 
 
 @admin.register(ClosedPeriod)
