+++ backend/samfundet/views.py
@@ -45,46 +45,35 @@ from .models.recruitment import (
 from .models.general import (
     Tag,
     User,
-    Menu,
     Gang,
-    Table,
     Venue,
     Image,
     Infobox,
     Profile,
-    Booking,
-    MenuItem,
     GangType,
     TextItem,
     KeyValue,
     BlogPost,
-    Reservation,
     Organization,
-    FoodCategory,
     Saksdokument,
     ClosedPeriod,
-    FoodPreference,
     UserPreference,
     InformationPage,
 )
 from .serializers import (
     TagSerializer,
     GangSerializer,
-    MenuSerializer,
     UserSerializer,
     ImageSerializer,
     EventSerializer,
-    TableSerializer,
     VenueSerializer,
     LoginSerializer,
     GroupSerializer,
     InfoboxSerializer,
     ProfileSerializer,
-    BookingSerializer,
     RegisterSerializer,
     TextItemSerializer,
     KeyValueSerializer,
-    MenuItemSerializer,
     GangTypeSerializer,
     BlogPostSerializer,
     InterviewSerializer,
@@ -92,14 +81,11 @@ from .serializers import (
     RecruitmentSerializer,
     SaksdokumentSerializer,
     OrganizationSerializer,
-    FoodCategorySerializer,
     ClosedPeriodSerializer,
     InterviewRoomSerializer,
-    FoodPreferenceSerializer,
     UserPreferenceSerializer,
     InformationPageSerializer,
     OccupiedtimeslotSerializer,
-    ReservationCheckSerializer,
     UserForRecruitmentSerializer,
     RecruitmentPositionSerializer,
     RecruitmentAdmissionForGangSerializer,
@@ -227,12 +213,6 @@ class IsClosedView(ListAPIView):
         )
 
 
-class BookingView(ModelViewSet):
-    permission_classes = (DjangoModelPermissionsOrAnonReadOnly, )
-    serializer_class = BookingSerializer
-    queryset = Booking.objects.all()
-
-
 class SaksdokumentView(ModelViewSet):
     permission_classes = (DjangoModelPermissionsOrAnonReadOnly, )
     serializer_class = SaksdokumentSerializer
@@ -273,66 +253,6 @@ class BlogPostView(ModelViewSet):
     permission_classes = (DjangoModelPermissionsOrAnonReadOnly, )
     serializer_class = BlogPostSerializer
     queryset = BlogPost.objects.all()
-
-
-# =============================== #
-#            Sulten               #
-# =============================== #
-
-
-class MenuView(ModelViewSet):
-    permission_classes = (DjangoModelPermissionsOrAnonReadOnly, )
-    serializer_class = MenuSerializer
-    queryset = Menu.objects.all()
-
-
-class MenuItemView(ModelViewSet):
-    permission_classes = (DjangoModelPermissionsOrAnonReadOnly, )
-    serializer_class = MenuItemSerializer
-    queryset = MenuItem.objects.all()
-
-
-class FoodCategoryView(ModelViewSet):
-    permission_classes = (DjangoModelPermissionsOrAnonReadOnly, )
-    serializer_class = FoodCategorySerializer
-    queryset = FoodCategory.objects.all()
-
-
-class FoodPreferenceView(ModelViewSet):
-    permission_classes = (DjangoModelPermissionsOrAnonReadOnly, )
-    serializer_class = FoodPreferenceSerializer
-    queryset = FoodPreference.objects.all()
-
-
-class TableView(ModelViewSet):
-    permission_classes = (DjangoModelPermissionsOrAnonReadOnly, )
-    serializer_class = TableSerializer
-    queryset = Table.objects.all()
-
-
-class ReservationCheckAvailabilityView(APIView):
-    permission_classes = [AllowAny]
-    serializer_class = ReservationCheckSerializer
-
-    def post(self, request: Request) -> Response:
-        serializer = self.serializer_class(data=request.data)
-        if serializer.is_valid():
-            if serializer.validated_data['reservation_date'] <= timezone.now().date():
-                return Response(
-                    {
-                        'error_nb': 'Reservasjoner må dessverre opprettes minst én dag i forveien.',
-                        'error_en': 'Unfortunately, reservations must be made at least one day in advance.'
-                    },
-                    status=status.HTTP_406_NOT_ACCEPTABLE
-                )
-            venue = self.request.query_params.get('venue', Venue.objects.get(slug='lyche').id)
-            available_tables = Reservation.fetch_available_times_for_date(
-                venue=venue,
-                seating=serializer.validated_data['guest_count'],
-                date=serializer.validated_data['reservation_date'],
-            )
-            return Response(available_tables, status=status.HTTP_200_OK)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
 # =============================== #
